restructured url fragments; started making views

+++ src/plugin_inbox.py
@@ -11,19 +11,19 @@ from minirpc import rpcmethod, RPCServable
 
 class Inbox(object) :
     @staticmethod
-    def add_to_inbox(blob, web) :
+    def add_to_inbox(user, web, blob) :
         blob_id = blob.id if isinstance(blob, models.Blob) else blob
         with models.DB :
-            models.DB.execute("insert into plugin_inbox (web_id, blob_id) values (?,?)", (web.id, blob_id))
+            models.DB.execute("insert into plugin_inbox (user_id, web_id, blob_id) values (?,?,?)", (user.id, web.id, blob_id))
     @staticmethod
-    def remove_from_inbox(blob, web) :
+    def remove_from_inbox(user, web, blob) :
         blob_id = blob.id if isinstance(blob, models.Blob) else blob
         with models.DB :
-            models.DB.execute("delete from plugin_inbox where web_id=? and blob_id=?", (web.id, blob_id))
+            models.DB.execute("delete from plugin_inbox where user_id=? and web_id=? and blob_id=?", (user.id, web.id, blob_id))
     @staticmethod
-    def get_inbox_uuids(web) :
+    def get_inbox_uuids(user, web) :
         web_id = web.id if isinstance(web, models.Web) else web
-        return [r["uuid"] for r in models.DB.execute("select uuid from blobs inner join plugin_inbox on blobs.id=plugin_inbox.blob_id where web_id=?", (web_id,))]
+        return [r["uuid"] for r in models.DB.execute("select uuid from blobs inner join plugin_inbox pi on blobs.id=pi.blob_id where pi.user_id=? and web_id=?", (user.id, web_id,))]
 
 @rpc_module("inbox")
 class InboxRPC(RPCServable) :
@@ -35,17 +35,18 @@ class InboxRPC(RPCServable) :
         if models.UserWebAccess.can_user_access(user, webid) :
 #            import relations
 #            return dict((i, relations.get_relations_for(webid, i)) for i in Inbox.get_inbox_uuids(webid))
-            return Inbox.get_inbox_uuids(webid)
+            return Inbox.get_inbox_uuids(user, webid)
         else :
             return None
 
 class InboxMessage(channel.Message) :
-    def __init__(self, blob, web, adding=True) :
-        self.blob = blob
+    def __init__(self, user, web, blob, adding=True) :
+        self.user = user
         self.web = web
+        self.blob = blob
         self.adding = adding
     def appropriate_for(self, user) :
-        return models.UserWebAccess.can_user_access(user, self.web)
+        return self.user.id == user.id
     def serialize(self) :
         import relations
         return {"type" : "InboxMessage",
@@ -62,8 +63,9 @@ def add_inbox_plugin(channels) :
         myMessages = []
         for m in newBlobMessages :
             for web in models.WebBlobAccess.get_webs_for_blob(m.blob) :
-                Inbox.add_to_inbox(m.blob, web)
-                myMessages.append(InboxMessage(m.blob, web, True))
+                for user in models.UserWebAccess.users_for_web(web) :
+                    Inbox.add_to_inbox(user, web, m.blob)
+                    myMessages.append(InboxMessage(user, web, m.blob, True))
         channels.broadcast(myMessages)
     channels.add_firehose_listener(inbox_channel_callback)
